Add OctoApp Plugin integration
Integrate OctoApp Plugin for Moonraker, enabling remote push notifications to mobile apps.
Plugin repo: https://github.com/crysxd/OctoApp-Plugin

- install OctoApp Plugin in "klipper install mode"
- add init script to start it as a service
- install Python dependencies using our normal chroot-helper
- add enable toggle config to extended.cfg

Known issues:
- Moonraker "host" parameter has to be changed from default "all" to some CIDR
  - OctoApp Plugin doesn't know how to parse "all"
- A lot of warnings in Moonraker/Fluidd/Mainsail because OctoApp insists on enabling updates

General notes:
- OctoApp Plugin is being extremely opinionated when configuring Moonraker via API
  - doesn't even expose overrides for its aggressive behavior (no env vars, config params, json params). Examples of problems caused by this:
    - can't tell octoapp to use different moonraker host, it insists on always reading it from Moonraker even though it doesn't know how to parse Moonraker's options properly
    - no config options to tell it to not enable updates
    - no options to tell it to not enable service supervisor nor polkit nor dbus
      - Moonraker has these disable by default because they don't exist on U1

Next steps:
- Try configure it in "Companion mode" (i.e. as if it's not running on the same server as moonraker).
  - This disables reading Moonraker config and appears to allow us to configure listen adress and keep Moonraker on both ipv4 and ipv6 (the "all" setting)
  - Hoping it allows us to also tell OctoApp to stop acting like it owns the whole world

Here are notifications it creates in Moonraker:

[update_manager]: Failed to load extension octoapp: [update_manager octoapp]: Invalid virtualenv at path /usr. Verify that the 'virtualenv' option is set to a valid virtualenv path.
Unparsed config option 'primary_branch: release' detected in section [update_manager octoapp]. This may be an option no longer available or could be the result of a module that failed to load. In the future this will result in a startup error.
Unparsed config option 'origin: https://github.com/crysxd/OctoApp-Plugin.git' detected in section [update_manager octoapp]. This may be an option no longer available or could be the result of a module that failed to load. In the future this will result in a startup error.
Unparsed config option 'requirements: requirements.txt' detected in section [update_manager octoapp]. This may be an option no longer available or could be the result of a module that failed to load. In the future this will result in a startup error.
Unparsed config option 'install_script: install.sh' detected in section [update_manager octoapp]. This may be an option no longer available or could be the result of a module that failed to load. In the future this will result in a startup error.
Unparsed config option 'managed_services:
octoapp' detected in section [update_manager octoapp]. This may be an option no longer available or could be the result of a module that failed to load. In the future this will result in a startup error.
Unable to find DBus PolKit Interface, this suggests PolKit is not installed on your OS.
Unable to initialize System Update Provider for distribution: buildroot